fix(core): refresh swagger docs for media and store menu_images (#190)
